backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pandas>=2.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
PyQt5>=5.15.0
matplotlib>=3.7.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
